+++ code/prop_data/atomic/get_event_triples.py
