+++ code/prop_data/atomic/paths.cfg
@@ -1,0 +1,21 @@
+[paths]
+atomic_tst = ../../../data/atomic/raw/v4_atomic_tst.csv
+atomic_trn = ../../../data/atomic/raw/v4_atomic_trn.csv
+atomic_dev = ../../../data/atomic/raw/v4_atomic_dev.csv
+atomic_raw = ../../../data/atomic/raw/v4_atomic_all_agg.csv
+atomic = ../../../data/atomic/prop/atomic.all
+atomic_event = ../../../data/atomic/prop/atomic_event.json
+atomic_rel = ../../../data/atomic/prop/atomic_relation.json
+atomic_graph = ../../../data/atomic/prop/atomic.graph
+atomic_sample_path = ../../../data/atomic/prop/sample_path_atomic.txt
+train_path = ../../../data/atomic/prop/train.txt
+dev_path = ../../../data/atomic/prop/dev.txt
+test_path = ../../../data/atomic/prop/test.txt
+train_relation_embedding_id= ../../../data/atomic/prop/train_relation_emb_id.json
+dev_relation_embedding_id= ../../../data/atomic/prop/dev_relation_emb_id.json
+test_relation_embedding_id= ../../../data/atomic/prop/test_relation_emb_id.json
+
+all_event_triples = ../../../data/atomic/event_triples/event_triples.json
+train_event_triples = ../../../data/atomic/event_triples/train_event_triples.txt
+dev_event_triples = ../../../data/atomic/event_triples/dev_event_triples.txt
+test_event_triples = ../../../data/atomic/event_triples/test_event_triples.txt
