+++ code/prop_data/empatheticdialogues/paths.cfg
@@ -1,0 +1,17 @@
+[paths]
+raw_tst = ../../../data/empatheticdialogues/raw/test.csv
+raw_trn = ../../../data/empatheticdialogues/raw/train.csv
+raw_dev = ../../../data/empatheticdialogues/raw/valid.csv
+prop_tst = ../../../data/empatheticdialogues/prop/ed_test.json
+prop_trn = ../../../data/empatheticdialogues/prop/ed_train.json
+prop_dev = ../../../data/empatheticdialogues/prop/ed_dev.json
+train_event_pairs = ../../../data/empatheticdialogues/prop/train_event_pairs.txt
+dev_event_pairs = ../../../data/empatheticdialogues/prop/dev_event_pairs.txt
+test_event_pairs = ../../../data/empatheticdialogues/prop/test_event_pairs.txt
+train_pred_relations = ../../../data/empatheticdialogues/prop/train_pred_relations.pkl
+dev_pred_relations = ../../../data/empatheticdialogues/prop/dev_pred_relations.pkl
+test_pred_relations = ../../../data/empatheticdialogues/prop/test_pred_relations.pkl
+
+train_relation_embedding_id= ../../../data/empatheticdialogues/prop/train_relation_emb_id.json
+dev_relation_embedding_id= ../../../data/empatheticdialogues/prop/dev_relation_emb_id.json
+test_relation_embedding_id= ../../../data/empatheticdialogues/prop/test_relation_emb_id.json
