--- data/atomic/paths.cfg
@@ -1,21 +1,0 @@
-[paths]
-atomic_tst = ./raw/v4_atomic_tst.csv
-atomic_trn = ./raw/v4_atomic_trn.csv
-atomic_dev = ./raw/v4_atomic_dev.csv
-atomic_raw = ./raw/v4_atomic_all_agg.csv
-atomic = ./prop/atomic.all
-atomic_event = ./prop/atomic_event.json
-atomic_rel = ./prop/atomic_relation.json
-atomic_graph = ./prop/atomic.graph
-atomic_sample_path = ./prop/sample_path_atomic.txt
-train_path = ./prop/train.txt
-dev_path = ./prop/dev.txt
-test_path = ./prop/test.txt
-train_relation_embedding_id= ./prop/train_relation_emb_id.json
-dev_relation_embedding_id= ./prop/dev_relation_emb_id.json
-test_relation_embedding_id= ./prop/test_relation_emb_id.json
-
-all_event_triples = ./event_triples/event_triples.json
-train_event_triples = ./event_triples/train_event_triples.txt
-dev_event_triples = ./event_triples/dev_event_triples.txt
-test_event_triples = ./event_triples/test_event_triples.txt
