--- data/empatheticdialogues/paths.cfg
@@ -1,17 +1,0 @@
-[paths]
-raw_tst = ./raw/test.csv
-raw_trn = ./raw/train.csv
-raw_dev = ./raw/valid.csv
-prop_tst = ./prop/ed_test.json
-prop_trn = ./prop/ed_train.json
-prop_dev = ./prop/ed_dev.json
-train_event_pairs = ./prop/train_event_pairs.txt
-dev_event_pairs = ./prop/dev_event_pairs.txt
-test_event_pairs = ./prop/test_event_pairs.txt
-train_pred_relations = ./prop/train_pred_relations.pkl
-dev_pred_relations = ./prop/dev_pred_relations.pkl
-test_pred_relations = ./prop/test_pred_relations.pkl
-
-train_relation_embedding_id= ./prop/train_relation_emb_id.json
-dev_relation_embedding_id= ./prop/dev_relation_emb_id.json
-test_relation_embedding_id= ./prop/test_relation_emb_id.json
