--- data/event_relation_classifier/run.sh
@@ -1,3 +1,0 @@
-export CUDA_VISIBLE_DEVICES=0,1,2,3
-
-python3 relation_classifier.py --warmup_ratio 0.1
