+++ requirements.txt
@@ -1,0 +1,16 @@
+transformers.egg==info
+regex==2020.10.15
+torch==1.8.1
+tqdm==4.59.0
+numpy==1.19.5
+pandas==1.1.3
+fairseq==0.10.2
+seaborn==0.11.0
+matplotlib==3.3.2
+nltk==3.5
+spacy==3.0.5
+apex==0.9.10dev
+pattern==3.6
+scikit_learn==1.0
+tensorflow==2.6.0
+transformers==4.11.3
